requirements.txt
# Homomorphic Encryption
tenseal>=0.3.14

# Data Structures
pybloom-live>=4.0.0

# HTTP Requests
requests>=2.31.0

# Numerical Operations
numpy>=1.24.0

# CLI Interface
click>=8.1.0

# API (Optional)
flask>=2.3.0
flask-cors>=4.0.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Performance & Benchmarking
matplotlib>=3.7.0
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
